// frontend/app/types.ts
// QUIZ TYPES
export type QuizData = {
  title: string;
  _id: string;
  questions: Question[];
};

export type Question = {
  text: string;
  imageUrl?: string;
  answers: string[];
  correctAnswer: number;
};

export type StudentAnswer = {
  quiz_id: string;
  choice: number;
  student_id: string;
  time_miliseconds: number;
};

export type LeaderBoardStatus = {
  quizName: string;
  rows: LeaderBoardRow[];
  final?: boolean;
};

export type LeaderBoardRow = {
  studentId: string;
  studenNick: string;
  score: number;
};

// managing types
export type ClassData = {
  students: Student[];
  cohortId: string;
  quizzesDone_Ids: string[];
};

export type Student = {
  name: string;
  email: string;
  nick: string;
  student_id: string;
};